Refactor: Add useSetDate hook
